Fix type of result

+++ GeneralAlgebraTeamLabProject/Lib/task7/discrete_logarithm.h
@@ -13,5 +13,4 @@ void print(const std::string_view rem, const M& mmap);*/
 
 LIB_EXPORT std::vector<int> discreteLogarithm(double a, double b, int p);
 
-LIB_EXPORT std::vector<long long> discreteLogarithm2(int base, int result, int module);
-
+LIB_EXPORT std::vector<unsigned long long> discreteLogarithm2(long base, long result, long module);
